+++ main.js
@@ -1,6 +1,6 @@
 import{getComments} from "./app.js"
 import{postComments} from "./app.js"
-
+import{formatDate} from "./assistant.js"
 
 const buttonElement = document.getElementById("add-form-button");
 const newCommentElement = document.getElementById("container");
@@ -52,23 +52,7 @@ initEventListeners();
 //формат даты
 const dateElement = document.getElementById("date");
 const myDate = new Date();
-const formatDate = (date) => {
-  let data = date.getDate();
-  let month = date.getMonth();
-  let hour = date.getHours();
-  let minute = date.getMinutes();
-
-  if (data < 10) {
-    data = "0" + data;
-  }
-  if (month < 10) {
-    month = "0" + (month + 1);
-  }
-  if (hour < 10) {
-    hour = "0" + hour;
-  }
-  return `${data}.${month}.${date.getFullYear().toString().substr(-2)} ${hour}:${minute}`;
-};
+formatDate();
 dateElement.textContent = formatDate(myDate);
 
 // массив данных
